Split AylythNoiseSettings into a few dedicated parts
